+++ frontend/package.json
@@ -1,36 +1,55 @@
 {
   "name": "@games_app/frontend",
-  "version": "1.0.0",
-  "scripts": {
-    "dev": "vite",
-    "build": "vite build",
-    "preview": "vite preview"
-  },
+  "version": "0.1.0",
+  "private": true,
   "dependencies": {
     "@games_app/shared": "*",
     "@hookform/resolvers": "^2.8.8",
+    "@testing-library/jest-dom": "^5.16.4",
+    "@testing-library/react": "^12.1.4",
+    "@testing-library/user-event": "^13.5.0",
     "antd": "^4.19.5",
     "axios": "^0.26.1",
     "dotenv": "^16.0.0",
     "http-status-codes": "^2.2.0",
-    "react": "^17.0.2",
+    "react": "17.0.2",
     "react-dom": "^17.0.2",
     "react-hook-form": "^7.29.0",
     "react-query": "^3.34.19",
-    "react-router-dom": "6",
+    "react-router-dom": "^6.3.0",
+    "react-scripts": "5.0.0",
     "recoil": "^0.7.0",
     "styled-components": "^5.3.5",
-    "vite-plugin-environment": "^1.1.0",
-    "vite-plugin-eslint": "^1.3.0",
+    "web-vitals": "^2.1.4",
     "yup": "^0.32.11"
   },
   "devDependencies": {
-    "@vitejs/plugin-react": "^1.0.7",
-    "@vitejs/plugin-react-refresh": "^1.3.6",
     "eslint": "^8.12.0",
     "eslint-plugin-react": "^7.29.4",
-    "eslint-plugin-react-hooks": "^4.4.0",
-    "vite": "^2.9.0",
-    "vite-plugin-svgr": "^1.1.0"
+    "eslint-plugin-react-hooks": "^4.4.0"
+  },
+  "scripts": {
+    "start": "react-scripts start",
+    "build": "react-scripts build",
+    "test": "react-scripts test",
+    "eject": "react-scripts eject"
+  },
+  "eslintConfig": {
+    "extends": [
+      "react-app",
+      "react-app/jest"
+    ]
+  },
+  "browserslist": {
+    "production": [
+      ">0.2%",
+      "not dead",
+      "not op_mini all"
+    ],
+    "development": [
+      "last 1 chrome version",
+      "last 1 firefox version",
+      "last 1 safari version"
+    ]
   }
 }
