+++ frontend/src/components/common/CustomHeading1/styles.js
