+++ frontend/src/components/common/CustomParagraph/styles.js
