+++ frontend/src/components/common/ProtectedRoute/index.js
