+++ frontend/src/components/pages/Auth/SignUp/index.js
@@ -24,8 +24,7 @@ const SignUp = () => {
     mode: 'onChange'
   });
 
-  const api = useAxios();
-
+  const { api } = useAxios();
   const { mutate, isLoading } = useMutation(authService.signUp(api), {
     onSuccess: ({ data }) => {
       navigate('/signin');
