+++ package.json
@@ -10,7 +10,7 @@
   "main": "index.js",
   "license": "MIT",
   "scripts": {
-    "frontend": "cd frontend && yarn dev",
+    "frontend": "cd frontend && yarn start",
     "backend": "cd backend && yarn dev",
     "dev": "concurrently \"yarn backend\" \"yarn frontend\""
   },
